Rename 'harbors' to 'harborList'
